+++ pssh/output.py
@@ -18,19 +18,27 @@
 
 """Output module of ParallelSSH"""
 
+from collections import namedtuple
 from os import linesep
 
 from . import logger
 
 
+HostOutputBuffers = namedtuple('Buffers', ['stdout', 'stderr'])
+BufferData = namedtuple('BufferData', ['reader', 'rw_buffer'])
+
+
 class HostOutput(object):
-    """Class to hold host output"""
+    """Host output"""
 
     __slots__ = ('host', 'channel', 'stdin',
-                 'client', 'exception', 'encoding', 'read_timeout')
+                 'client', 'exception', 'encoding', 'read_timeout',
+                 'buffers',
+                 )
 
     def __init__(self, host, channel, stdin,
-                 client, exception=None, encoding='utf-8', read_timeout=None):
+                 client, exception=None, encoding='utf-8', read_timeout=None,
+                 buffers=None):
         """
         :param host: Host name output is for
         :type host: str
@@ -54,13 +62,14 @@ class HostOutput(object):
         self.exception = exception
         self.encoding = encoding
         self.read_timeout = read_timeout
+        self.buffers = buffers
 
     @property
     def stdout(self):
         if not self.client:
             return
         _stdout = self.client.read_output_buffer(
-            self.client.read_output(self.channel, timeout=self.read_timeout),
+            self.client.read_output(self.buffers.stdout.rw_buffer, timeout=self.read_timeout),
             encoding=self.encoding)
         return _stdout
 
@@ -69,7 +78,7 @@ class HostOutput(object):
         if not self.client:
             return
         _stderr = self.client.read_output_buffer(
-            self.client.read_stderr(self.channel, timeout=self.read_timeout),
+            self.client.read_stderr(self.buffers.stderr.rw_buffer, timeout=self.read_timeout),
             encoding=self.encoding,
             prefix='\t[err]')
         return _stderr
@@ -84,16 +93,19 @@ class HostOutput(object):
             logger.error("Error getting exit status - %s", ex)
 
     def __repr__(self):
-        return "{linesep}\thost={host}{linesep}" \
+        return "\thost={host}{linesep}" \
             "\texit_code={exit_code}{linesep}" \
-            "\t{linesep}\tchannel={channel}{linesep}" \
+            "\tchannel={channel}{linesep}" \
             "\tstdout={stdout}{linesep}\tstderr={stderr}{linesep}" \
             "\tstdin={stdin}{linesep}" \
-            "\texception={exception}{linesep}".format(
+            "\texception={exception}{linesep}" \
+            "\tencoding={encoding}{linesep}" \
+            "\tread_timeout={read_timeout}".format(
                 host=self.host, channel=self.channel,
                 stdout=self.stdout, stdin=self.stdin, stderr=self.stderr,
                 exception=self.exception, linesep=linesep,
-                exit_code=self.exit_code)
+                exit_code=self.exit_code, encoding=self.encoding, read_timeout=self.read_timeout,
+            )
 
     def __str__(self):
         return self.__repr__()